Restructure app based on feature/server branch
