+++ .sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path');
+
+module.exports = {
+    'config': path.resolve('server/db/config', 'config.json'),
+    'migrations-path': path.resolve('server/db', 'migrations'),
+    'models-path': path.resolve('server', 'models'),
+    'seeders-path': path.resolve('server', 'seeds')
+};
