+++ package.json
@@ -7,7 +7,7 @@
     "test": "mocha --require babel-register --reporter spec || exit 0",
     "start": "nodemon dist/app",
     "clean": "rm -rf dist",
-    "build": "npm run clean && mkdir dist && babel src -s -d dist",
+    "build": "npm run clean && mkdir dist && babel server -s -d dist",
     "build:prod": "npm run build && node bin/prod"
   },
   "keywords": [],
