--- server/.sequelizerc
@@ -1,8 +1,0 @@
-const path = require('path');
-
-module.exports = {
-    'config': path.resolve('src/db/config', 'config.json'),
-    'migrations-path': path.resolve('src/db', 'migrations'),
-    'models-path': path.resolve('src', 'models'),
-    'seeders-path': path.resolve('src', 'seeds')
-};
