+++ server/models/index.js
@@ -5,7 +5,7 @@ import { Sequelize } from 'sequelize';
 
 const basename = path.basename(module.filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require('../../src/db/config/config.json')[env];
+const config = require('../../server/db/config/config.json')[env];
 const db = {};
 
 let sequelize;
